requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Login>=0.6,<1.0
Werkzeug>=3.0,<4.0
Flask-WTF>=1.2,<2.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
